Website Navigational Template
I have created the README alongside a template to build our website off of.

in the master style sheet, the `:root` element contains information such as thematic colors and the width of the navbar.

using `:root` will allow us to more easily change the theme of the website when that comes around.

The navbar is well accessible, supporting use of tab navigation. The navbar collapses when the screen gets small, and will open when either the mouse gets close to the left or the user uses tab navigation.