Page indices, User Preference framework
I've added a system which works using `localStorage` to store thematic preferences in the browser.

If we want to export this theme preference data too, I've implemented a helper function to return all user-modified preference data. it is called `QueryAllPrefData()` and returns an array of objects. each object has at least a 'name' and a 'data' field, and some which are specific page preferences have an additiona 'page' field for the page index of the preference.

Currently, it fully works. I have implemented all of the functions and framework to easily add new user-customizable variables, and save that information to the `localStorage`

If you want to test it out, use the command `SetPref(key, val)` in the console. The lists of editable properties are shown in `userprefs.js`.

to clear user browser prefs, i have added a function `ClearUserPrefs()` which does exactly that.