JSDoc Complete, Minor Refactoring
I have completed the UserThemePrefs library well enough and documented it that I hopefully don't have to do much in this file anymore.

with that being said, I think I'm really happy with the result! I modified the GitHub to include information about it.

If you're interested in adding customization menus for your little box thingies, that's the entire reason I made this library