Class Encapsulation of UTP
UserThemePrefs is now comprised of static functions as well as instance functions for UTP objects.

This was done to make calls to common functions more convenient, as well as to hopefully make this whole UTP thing more intuitive.

I'm not sure if it matters for the purpose of our project, but now UserThemePrefs is a namespace.